backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
bcrypt
certifi
python-jose[cryptography]
